Renamed SystemManager to SystemsManager for clarity
Changes made in the files targeted towards renaming the SystemManager to SystemsManager which better represents its functionalities. This change has been propagated throughout the codebase with all instances of SystemManager being updated to SystemsManager. Doing so helps improve the readability and comprehension of the codebase making for efficient code understanding for collaborators or newcomers to the project.