Update GameLoop and EditorGameLoop for better handling of windows
Refactored the GameLoop and EditorGameLoop classes to improve window handling. The constructors now take a shared pointer to a Window object, allowing it to be stored within the class. The loop() methods no longer take a Window object as a parameter, significantly simplifying the method signature. Moreover, all float data types related to frame timelocking and interpolation calculation changed with double for better precision, and more correct semantically. EditorGui instance is now created in the constructor, it becomes possible due to moving the window into a class field.

This refactoring provides a more intuitive and safer way of keeping track of the Window object. It also significantly simplifies the function signatures and makes the code easier to read and understand.