"Updated DebugRenderSystem and TransformComponent matrix transformations
In DebugRenderSystem changed creating debugGridTransform from a zero vector to a unit vector to ensure visible grid in debug mode. Invoked updateTransformMatrix to apply the new transform right away.
In TransformComponent.cpp, updateTransformMatrix was modified to only perform transformations if component position or state is marked as 'dirty'. This reduces needless calculations when state is unchanged. Added a call to mark as 'dirty' in setPosition to ensure transformations applied if position is changed."