Adjust camera settings for isometric view
The camera settings have been adjusted to support both isometric and perspective views. An isometric boolean flag was introduced to the CameraComponent. This is to accommodate games that might require a non-perspective, isometric view. Changes include, refactor of rotate() method into yaw() and pitch() methods for clearer code. The field of view in the EditorSystemsManager was also narrowed from 45.0f to 25.0f to suit the isometric view. These changes will make it more convenient for developers working on games needing an isometric view in our engine.