Added dynamic grid and related functionalities
This commit includes significant enhancements to the core::Grid class. A dynamic, two-dimensional grid has been implemented in the Grid.cpp along with a destructor to manage memory properly. The grid is composed of cells, which are represented by a nested struct within the core::Grid class. For managing interactions with cells, a new method 'checkIfPointInCell' has been added which takes a position as vec2 and returns a shared pointer to the cell containing the point, or nullptr if no such cell is found.

This new feature will help in managing game entities and their interactions spatially, improving efficiency and potential for more complex gameplay mechanisms.