Refactor EditorGUI to inherit from IGUI interface
EditorGUI is now a subclass of the IGUI interface. With this change, EditorGUI's method 'processEvent' is declared as override and the argument is changed from a pointer to a const reference for enhanced safety. Consequently, the internal method call 'ImGui_ImplSDL3_ProcessEvent' is adjusted to take the address of 'event'. Moreover, the required include statement and IGUI.h file are added to EditorGUI.h and CMakeLists.txt respectively. This reinforces the use of interfaces and enhances type safety in the codebase.