Refactor naming for mouse button states
The name of the states that represent the status of the mouse buttons were changed. The term "Mouse" was replaced with "Button" to make the code more intuitive and easier to understand. It also aligns better with the syntax used for the rest of the button state variables. Furthermore, it improves the readability of the code for future maintenance and development.