Refactor build system and rename project to Hexen
This commit restructures the CMake system by consolidating and streamlining includes and linkers. It also renames the project from 'main' to 'Hexen'. These changes are made to improve the build system organization, offer better project identification and to enable easier future modifications. The BUILD_ANDROID and BUILD_LIB options are also modified to allow either static or shared library configurations, with the addition of an input system and GUI components to the engine. Furthermore, the editor's build system is updated to reference the reorganized engine build system. Overall, these changes improve the build system's flexibility and the project's structure.