Update SDL initialization to support all subsystems
This commit updates the call to 'SDL_Init' in 'Window::initSDL()' from 'SDL_INIT_VIDEO' to 'SDL_INIT_EVERYTHING'. This change is necessary to include support for gamepad input, as we're introducing a new 'Gamepad' class in the 'core::input' namespace. With the updated SDL initialization, the engine will now be able to handle and process gamepad inputs, providing a more interactive option for users.