Add Gamepad support
This commit integrates Gamepad support into the InputSystem class. This includes processing Gamepad inputs similarly to Mouse and Keyboard inputs. Gamepads are added or removed dynamically when connected or disconnected. This allows for a more inclusive and diverse user experience, as well as enriches the player's interaction with the game. Also, a gamepad's thumbstick positions are now output in the console, providing debug information for developers.