Add Keyboard support and functionality
This commit introduces a new class, Keyboard, into the InputSystem. The Keyboard class includes methods to process keyboard inputs and check the state of a specific key. An extensive enum Key is also created for all possible key values. This will allow developers to easily check the state of any key on the keyboard, enabling efficient creation of keyboard interactions in the game. Improvements have also been made to the CMakeLists.txt for seamless integration of the new files.