Refactored mouse input handling in InputSystem
This commit refactors handling of mouse input in the InputSystem. A function "processMouseInput" was added to GeometrySystem which processes any mouse input and corresponding callback functions. New fields were added to the Mouse class to track the movement of the mouse along X and Y axes, and the wheel position. Furthermore, method "getWheelPosition" was implemented in the Mouse class to get the current position of the mouse wheel. Changes are motivated by the need to enhance mouse input handling capability in our engine.