Change errMsg declaration in ComponentNotFound exception
Modify the exceptions 'ComponentNotFound' to decouple the error message declaration from the 'what()' function call. A 'message' string attribute is added to store the actual error message during the object's initialization period. This change is to enhance the flexibility of generated error msgs, reduce redundancy, and achieve better exception management.