Add MappingNotFound exception to engine exceptions
Added a new exception, 'MappingNotFound' under the hexen::engine::exceptions namespace. This exception is thrown when a specified axis/action mapping is not found. The exception includes a constructor storing the mapping name & a what() function to relay a message. This enhancement improves the debugging process by providing more specific error messages for mapping-related issues.