Introduced InputSystem in SystemsManager and EditorSystemsManager
The main system now includes a new system, InputSystem to handle inputs from the window. The processInput method now uses this system instead of processing SDL events directly in SystemsManager itself. The InputSystem is shared as a protected attribute, allowing child classes to use it. Also the usage of inputSystem has been reflected in EditorSystemsManager where the initialization and usage of inputSystem have been updated. Now this system is initialized in SystemsManager. This change focuses on increasing modularity and enhancing system-specific responsibility handling.