Replace manual creation of TransformComponent with system call
In this commit, the manual creation of TransformComponent in the SceneEntity constructor has been replaced with a call to RenderSystem's createTransformComponent. This is done to increase consistency and to allow the system to handle TransformComponent creation, as previously signaled by the TODO comment in the code. This simplifies the code and makes it more maintainable.