Remove default isometric flag from CameraComponent constructor
In this commit, the isometric flag default value from the constructor of the CameraComponent is removed. Previously, constructors for these classes had an optional isometric parameter that would default to true if no argument was given.

This change was made because not all cameras in our engine are isometric by default, making this assumption misleading. This might steer developers into potential bugs, especially newcomers to the project unaware of this camera’s behavior. Therefore, it's better to make this explicit in the constructor call.

With this modification, any new instance of a CameraComponent will require explicit setting of this property, hence leading to better control of camera configurations. By requiring this specification, it ensures that developer intent is evident in the code. This clarity helps to avoid assumptions in the game engine that can cause future headaches and bugs when building new features or games.