Replace default camera with custom EditorCameraComponent
This commit introduces the use of a custom EditorCameraComponent in EditorSystemsManager. The motivation behind this is to allow the editor to have more customized control over the camera component.

This change includes the addition of an updateCameras function in DebugRenderSystem which assures that the cameras are updated correctly in the context of the editor. To match this change, we've also defined an explicit constructor and updated the update function in EditorCameraComponent.

The shift to a custom camera within the editor paves the way for unique editor-related camera behaviours, potentially enhancing the usability of the game engine’s editing capabilities.