Update SceneEntity's comments and improve findNode implementation
Expanded, detailed comments were added to the SceneEntity class in order to improve understanding of this class and its methods. In addition, searchNode method was renamed to findNode to convey clearer intent. Its implementation was also updated to refer to "foundedNodes" instead of "findedNodes" for correct English grammar. Furthermore, a redundant empty line was removed. This refactoring is essential to maintain clean, understandable and grammatically-correct code, making it easier for other developers to read and contribute.