Refactor core engine and separate client application
In this commit, the engine's core application code has been refactored and separated into a stand-alone client application. This includes restructuring of files, adjusting includes, and modifying creation of certain elements like TransformComponent and Application.

The changes were made to improve the organization of the codebase, making it easier to navigate and manage. Furthermore, separating the client application from the core engine better adheres to the separation of concerns principle, improving maintainability and scalability.

An option for building for Android was also added to the CMake configuration, paving the way for cross-platform compatibility. Some files were renamed to better fit their contents and a new Application.rc and corresponding CMakeLists.txt were added for the client application project.

The CMakeLists.txt file of the engine core has also been changed, removing the target sources related to Application as it is now a part of the client application. The .gitignore file was updated to ignore additional directories.

Additionally, a range of files were renamed and moved to fit with the new project structure, particularly several files related to Android app configuration and asset management.

Overall, these changes contribute to a cleaner, more organized codebase that is easier to develop, maintain and deploy across various platforms.