Update Android app's minimum and target SDK version
The minimum SDK version was raised from 16 to 21 and the target SDK version was removed. This is to ensure that the app is compatible with more recent Android versions and APIs. The CMake build script was also updated reflecting these changes, along with some adjustments for the game name and path to the CMakeLists.txt file. These changes improve the app's compatibility and adaptability for future development.