Update thread handling for Android compatibility
This commit addresses issues with thread handling in the Android environment. We've replaced instances of FTL_DEBUG with HEXEN_DEBUG in Config.hpp and added conditions to check for an Android build in both the Config.hpp and Thread.cpp files to adjust the handling of threading appropriately. Furthermore, we've changed the pthread_t typedef to better support Android platforms within the Thread.hpp file. We've taken these measures to ensure proper execution of the existing code on Android systems, also enhancing the cross-platform compatibility of our threading operations.