Reorder CMake project definition for clarity
The project declaration in CMakeLists.txt was reordered based on the EDITOR_NAME variable. Before, the project name was defined before the variable that it depends on, which may have caused confusion. Now the EDITOR_NAME variable is set before it's used in the project declaration making the code less prone to errors and easier to understand.