Refactor argument name in ConstIterator's copy constructor
The argument name for the copy constructor of class ConstIterator has been changed from "hashTable" to "iterator". The previous name could lead to confusion as it was suggesting a whole hashtable is being copied while we are copying an iterator. This rename will make it clearer to read and maintain the code.