Update forEach to use parallel execution
Switched the forEach implementation in hexen::engine::core::Scene to utilize std::for_each with std::execution::par to allow for parallel execution for efficiency improvement. Consequently, added the necessary includes for algorithm, utility, and execution. Included a warning in the method's documentation to recommend the use of synchronization primitives to avoid data racing and deadlocks when passing functions to forEach method.