Update ComponentContainer and correct CMakeLists
Updated ComponentContainer's operators and access patterns to return references and pointers, enhancing efficiency and control. Also, the iterator is now a friend class, promoting encapsulation. In CMakeLists.txt, ComponentContainer.cpp was removed as it is no longer used in the project. These changes streamline data access and adhere to best programming practices.