Refactor code in EditorSystemsManager and ComponentContainer files
The update in EditorSystemsManager.cpp simplifies the task creation code and calls the RenderSystem directly instead of wrapping it in a lambda function. The update in ComponentContainer.hpp include an explicit default constructor to ensure proper object creation and changes the iterator_category to a forward_iterator. These changes improve the code efficiency and flow, and align with standard programming practices. Additionally, the .gitignore file has been updated to ignore more build files.