Refactor graphics classes for OpenGL specificity
Streamlined the structure of graphics classes in the engine to align more contextually with OpenGL specific naming and hierarchy conventions. Renamed Classes and files with 'GL' prefix to distinctively highlight their tertiary relation to OpenGL. This aids in the clarity of the codebase and makes it more user friendly for those familiar with OpenGL. For instance, a VertexBufferObject is now GLVertexBufferObject and stored in 'engine/src/graphics/buffers/GL' instead of 'engine/src/core/graphics'. Changes applied uniformly to all related classes, files and their usage. Also, updated references and documentation accordingly for consistency.