Implement dynamic choice of RenderAPI
Refactored the creation of the unique pointer of RenderContext to support the dynamic choice of RenderAPI. Addition of the set method to 'currenApi' will enable runtime selection of the rendering API based on user settings. This was achieved by updating the create() method and through the implementation of a switch case strategy.

The dynamic choice of RenderAPI leads to better flexibility and adaptability of the RenderContext. The strategy can now handle the inclusion of new rendering APIs without codebase disruption. Also, there are enhancements to code readability and maintainability.