Rearrange render API switch cases for better logic flow
Rearranged the order of cases in the switch statement that handles the selected render API. The change is intended to provide an immediate feedback in case the NO_API scenario is encountered, by placing it as the first case of the switch. This reduces the overhead of unnecessary checks for other cases when no API has been selected or if the device does not support graphics.