Update CMakeLists to reflect correct import paths
Adjusted CMakeLists.txt in the graphics engine to accurately reflect the correct import paths for the parent directory containing the core components of the engine. This change makes the RenderContext include statements public. This ensures that the correct relative paths are used irrespective of the location where the build process is initiated, hence increasing the portability of the project.