Added support of FrameBuffer and improved name conventions
Refactored 'GLFrameBufferObject' to 'GLFrameBuffer', 'GLRenderBufferObject' to 'GLRenderBuffer', added explicit constructors with input parameters and improved the structure of the Buffer classes to make them more readable and understandable. Also updated CMakeLists.txt file to include new GLFrameBuffer and GLRenderBuffer classes. These changes are made to add support of FrameBuffer in OpenGL rendering, and align the class names with standard naming conventions. Added FrameBuffer class definition to Buffer.hpp to provide an abstract base for new FrameBuffer implementations and improve the extendability of the code.