Remove redundant Graphics.hpp file and enhance GLVertexBuffer
Deleted the unnecessary Graphics.hpp file from the graphics/buffers/GL directory which was not providing any additional functionality or encapsulation. This was done to improve the maintainability of the code by keeping it concise and free of clutter.

Enhanced the GLVertexBuffer class by adding functions setLayout and getLayout for setting and retrieving the new layout of the buffer. Also, the GLVertexBuffer class was equipped with an additional attribute, bufferLayout which stores the current layout of the buffer. This was done to make the class more versatile and self-contained resulting in better processing of buffer data.

Also added buffers in Buffers.cpp enriching its capacity to handle buffer data. Modifications in Buffers.hpp involves adding enumeration for ShaderDataType and structure BufferElement and class BufferLayout which deals with buffer layout data. The code refactoring was done to handle buffer data more efficiently, by introducing a more structured approach to the buffer's data layout.