Refactor ShaderProgram and RenderContext for better encapsulation
Renamed the ShaderProgram file and class to GLShaderProgram to highlight its specific implementation for the OpenGL environment. Made changes to the GLShaderProgram and RenderContext constructors to better manage shader files and the SDL window, respectively, enhancing flexibility and maintainability. Modified include directives to use quotes instead of angle brackets for local files.