Implement FramebufferCommand class in render_commands
Implemented FramebufferCommand class to process framebuffer commands. This class includes FrameBuffer specification methods allowing the execution, preparation, and completion of framebuffer commands. Added getColorAttachment method to retrieve the color attachment of the framebuffer. This addition facilitates smoother processing of framebuffer commands and improves the performance of the graphic rendering pipeline.