Implement RenderPipeline for graphic commands
This commit introduces the RenderPipeline class providing a pipeline for rendering operations. This class improves graphics rendering performance by organizing rendering commands into a queue and processing them in order.

Two main functions added, executeCommandNow and addCommandToQueue, allow instantaneous execution of rendering commands and adding them to the queue respectively. It also includes getCommand, removeCommandFromQueue, executeCommands, prepareCommands, and finishCommands methods for manipulating and executing render commands in an efficient manner.

Updated the CMakeLists.txt file to link the appropriate libraries for the addition of the RenderPipeline to the Hexen engine.